+++ services/storage/src/simcore_service_storage/modules/datcore_adapter/datcore_adapter.py
@@ -1,8 +1,6 @@
 import logging
-from collections.abc import Callable
-from math import ceil
 from pathlib import Path
-from typing import Any, TypeAlias, TypeVar, cast
+from typing import Any, TypeAlias, cast
 
 import httpx
 from fastapi import FastAPI
@@ -36,106 +34,10 @@ from ...models import (
     PathMetaData,
     TotalNumber,
 )
-from .datcore_adapter_exceptions import (
-    DatcoreAdapterClientError,
-    DatcoreAdapterError,
-    DatcoreAdapterTimeoutError,
-)
+from .datcore_adapter_client_utils import request, retrieve_all_pages
+from .datcore_adapter_exceptions import DatcoreAdapterError
 
 _logger = logging.getLogger(__file__)
-
-
-class _DatcoreAdapterResponseError(DatcoreAdapterError):
-    """Basic exception for response errors"""
-
-    def __init__(self, status: int, reason: str) -> None:
-        self.status = status
-        self.reason = reason
-        super().__init__(
-            msg=f"forwarded call failed with status {status}, reason {reason}"
-        )
-
-
-async def _request(
-    app: FastAPI,
-    api_key: str,
-    api_secret: str,
-    method: str,
-    path: str,
-    *,
-    json: dict[str, Any] | None = None,
-    params: dict[str, Any] | None = None,
-    **request_kwargs,
-) -> dict[str, Any] | list[dict[str, Any]]:
-    datcore_adapter_settings = get_application_settings(app).DATCORE_ADAPTER
-    url = datcore_adapter_settings.endpoint + path
-    session = get_client_session(app)
-
-    try:
-        if request_kwargs is None:
-            request_kwargs = {}
-        response = await session.request(
-            method.upper(),
-            url,
-            headers={
-                "x-datcore-api-key": api_key,
-                "x-datcore-api-secret": api_secret,
-            },
-            json=json,
-            params=params,
-            **request_kwargs,
-        )
-        response.raise_for_status()
-        response_data = response.json()
-        assert isinstance(response_data, dict | list)  # nosec
-        return response_data
-
-    except httpx.HTTPStatusError as exc:
-        raise _DatcoreAdapterResponseError(
-            status=exc.response.status_code, reason=f"{exc}"
-        ) from exc
-
-    except TimeoutError as exc:
-        msg = f"datcore-adapter server timed-out: {exc}"
-        raise DatcoreAdapterTimeoutError(msg) from exc
-
-    except httpx.RequestError as exc:
-        msg = f"unexpected request error: {exc}"
-        raise DatcoreAdapterClientError(msg) from exc
-
-
-_T = TypeVar("_T")
-
-
-async def _retrieve_all_pages(
-    app: FastAPI,
-    api_key: str,
-    api_secret: str,
-    method: str,
-    path: str,
-    return_type_creator: Callable[..., _T],
-) -> list[_T]:
-    page = 1
-    objs = []
-    while (
-        response := cast(
-            dict[str, Any],
-            await _request(
-                app, api_key, api_secret, method, path, params={"page": page}
-            ),
-        )
-    ) and response.get("items"):
-        _logger.debug(
-            "called %s [%d/%d], received %d objects",
-            path,
-            page,
-            ceil(response.get("total", -1) / response.get("size", 1)),
-            len(response.get("items", [])),
-        )
-
-        objs += [return_type_creator(d) for d in response.get("items", [])]
-        page += 1
-    return objs
 
 
 async def check_service_health(app: FastAPI) -> bool:
@@ -156,7 +58,7 @@ async def check_user_can_connect(app: FastAPI, api_key: str, api_secret: str) ->
         return False
 
     try:
-        await _request(app, api_key, api_secret, "GET", "/user/profile")
+        await request(app, api_key, api_secret, "GET", "/user/profile")
         return True
     except DatcoreAdapterError:
         return False
@@ -200,7 +102,7 @@ async def list_all_files_metadatas_in_dataset(
 ) -> list[FileMetaData]:
     all_files: list[dict[str, Any]] = cast(
         list[dict[str, Any]],
-        await _request(
+        await request(
             app,
             api_key,
             api_secret,
@@ -269,7 +171,7 @@ async def list_top_level_objects_in_dataset(
     limit: NonNegativeInt,
 ) -> tuple[list[PathMetaData], GenericCursor | None, TotalNumber]:
     page, size = _init_pagination(cursor, limit)
-    response = await _request(
+    response = await request(
         app,
         api_key,
         api_secret,
@@ -277,6 +179,7 @@ async def list_top_level_objects_in_dataset(
         f"/datasets/{dataset_id}/files",
         params={"size": size, "page": page},
     )
+    assert isinstance(response, dict)  # nosec
     file_metadata_page = Page[DatCoreFileMetaData](**response)
     entries = file_metadata_page.items
     total = file_metadata_page.total
@@ -335,7 +238,7 @@ async def list_top_level_objects_in_collection(
     limit: NonNegativeInt,
 ) -> tuple[list[PathMetaData], GenericCursor | None, TotalNumber]:
     page, size = _init_pagination(cursor, limit)
-    response = await _request(
+    response = await request(
         app,
         api_key,
         api_secret,
@@ -343,6 +246,7 @@ async def list_top_level_objects_in_collection(
         f"/datasets/{dataset_id}/files/{collection_id}",
         params={"size": size, "page": page},
     )
+    assert isinstance(response, dict)  # nosec
     file_metadata_page = Page[DatCoreFileMetaData](**response)
     entries = file_metadata_page.items
     total = file_metadata_page.total
@@ -442,7 +346,7 @@ async def get_package_file_as_path(
 async def list_all_datasets(
     app: FastAPI, api_key: str, api_secret: str
 ) -> list[DatasetMetaData]:
-    all_datasets: list[DatasetMetaData] = await _retrieve_all_pages(
+    all_datasets: list[DatasetMetaData] = await retrieve_all_pages(
         app,
         api_key,
         api_secret,
@@ -464,7 +368,7 @@ async def list_datasets(
 ) -> tuple[list[DatasetMetaData], GenericCursor | None, TotalNumber]:
     page, size = _init_pagination(cursor, limit)
 
-    response = await _request(
+    response = await request(
         app,
         api_key,
         api_secret,
@@ -472,6 +376,7 @@ async def list_datasets(
         "/datasets",
         params={"size": size, "page": page},
     )
+    assert isinstance(response, dict)  # nosec
     datasets_page = Page[DatCoreDatasetMetaData](**response)
     datasets = datasets_page.items
     total = datasets_page.total
@@ -494,7 +399,7 @@ async def get_file_download_presigned_link(
 ) -> AnyUrl:
     file_download_data = cast(
         dict[str, Any],
-        await _request(app, api_key, api_secret, "GET", f"/files/{file_id}"),
+        await request(app, api_key, api_secret, "GET", f"/files/{file_id}"),
     )
     url: AnyUrl = TypeAdapter(AnyUrl).validate_python(file_download_data["link"])
     return url
@@ -504,7 +409,7 @@ async def get_package_files(
     app: FastAPI, *, api_key: str, api_secret: str, package_id: str
 ) -> list[PackageMetaData]:
     return TypeAdapter(list[PackageMetaData]).validate_python(
-        await _request(
+        await request(
             app,
             api_key,
             api_secret,
@@ -517,4 +422,4 @@ async def get_package_files(
 async def delete_file(
     app: FastAPI, api_key: str, api_secret: str, file_id: str
 ) -> None:
-    await _request(app, api_key, api_secret, "DELETE", f"/files/{file_id}")
+    await request(app, api_key, api_secret, "DELETE", f"/files/{file_id}")
